clean circular deps in telemetry
